Use `atom/highlights` for Code Splitting!
This is really fancy, as it eliminates the need for groc's complicated
separator (one of the few files I just copy&pasted).

So far, this works quite well for JS and Coffee files, but still depends
on the language definition in `language.coffee` for comment-only files
and removing leading characters from block comments.